fix: image upload not working when published
