+++ src/core/domain/dtos/__init__.py
@@ -1,4 +1,10 @@
-from .file import File, SignedUrl
+from .file import (
+    File,
+    SignedUploadUrlRequest,
+    SignedUrl,
+    SignedUrlRequest,
+    SignedUrlsRequest,
+)
 from .profile import PublishingOptionsUpdate, UpdateProfile
 from .token import AccessResponse, RefreshRequest, TokensResponse
 from .user import (
